Backend(Python)/requirements.txt
fastapi
uvicorn
motor
python-dotenv
google-genai
pydantic
